fix: route app, login, main pages
