+++ src/App.tsx
@@ -1,23 +1,18 @@
 import React from 'react';
-import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom';
-import './scss/main.scss';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { ReactNotifications } from 'react-notifications-component';
-import Home from './pages/main';
+import Login from './pages/Login/Login';
 import Error from './pages/error';
-import Login from './pages/Login/login';
+import Main from './pages/Main/Main';
 import Moderation from './pages/Moderation/Moderation';
 
 function App() {
   return (
     <Router>
       <ReactNotifications />
-
-      <nav>
-        <Link to="/" />
-      </nav>
       <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/Login/*" element={<Login />} />
+        <Route path="/login" element={<Login />} />
+        <Route path="/*" element={<Main />} />
         <Route path="*" element={<Error />} />
         <Route path="/Moderation" element={<Moderation />} />
       </Routes>
